requirements.txt
firecrawl-py>=4.19.0
supabase>=2.28.0
python-dotenv>=1.1.0
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
requests>=2.32.0
